next() python3 and refactor sql
